component-ize Nav
