+++ src/components/navigation.js
@@ -1,5 +1,5 @@
-import React from 'react'
-import Link from 'gatsby-link'
+import React from 'react';
+import Link from 'gatsby-link';
 import {
   Navbar,
   NavbarBrand,
@@ -9,68 +9,23 @@ import {
   NavbarLink,
   NavbarMenu,
   Title,
-  withHelpersModifiers,
-} from 'bloomer'
-import { secondaryFont, serifFont } from '../utils/fonts'
+} from 'bloomer';
+import { secondaryFont, serifFont } from '../utils/fonts';
 import {
   colorBrownDark,
   colorGreen,
   colorGreenDark,
-} from '../utils/theme-variables'
-
-const linkStyles = {
-  color: colorBrownDark,
-  borderBottom: `2px solid transparent`,
-}
-const activeLinkStyles = {
-  color: colorBrownDark,
-  borderBottom: `2px solid ${colorGreen}`,
-}
-
-const StyledNavbarItem = ({ to, text }) => (
-  <NavbarItem>
-    <Link
-      style={linkStyles}
-      activeStyle={activeLinkStyles}
-      to={`/${to || text.toLowerCase()}`}
-    >
-      {text}
-    </Link>
-  </NavbarItem>
-)
-
-const BrandItem = ({ title }) => (
-  <Title>
-    <Link
-      to="/"
-      css={{
-        ...secondaryFont,
-        color: colorGreen,
-        textTransform: 'none',
-        '&:hover': {
-          color: colorGreenDark,
-        },
-      }}
-    >
-      {title}
-    </Link>
-  </Title>
-)
+} from '../utils/theme-variables';
+import DesktopBrand from './nav/desktop-brand';
+import BrandItem from './nav/brand-item';
+import StyledNavbarItem from './nav/styled-navbar-item';
+import ProductsDropdownItem from './nav/products-dropdown-item';
+import TabletNavbarBrand from './nav/tablet-navbar-brand';
 
 const Navigation = ({ onMenuClick, isOpen, categoryNodes, title }) => {
   return (
     <div>
-      <div
-        className="is-hidden-touch"
-        css={{
-          width: '100%',
-          marginBottom: `0.5rem`,
-          textAlign: 'center',
-          paddingTop: '1.5em',
-        }}
-      >
-        <BrandItem title={title} />
-      </div>
+      <DesktopBrand title={title} />
       <Navbar
         style={{
           ...serifFont,
@@ -78,17 +33,7 @@ const Navigation = ({ onMenuClick, isOpen, categoryNodes, title }) => {
           background: 'transparent',
         }}
       >
-        <NavbarBrand isHidden="desktop">
-          <NavbarBurger
-            style={{ marginRight: 'auto', marginLeft: 'inherit' }}
-            isActive={isOpen}
-            onClick={onMenuClick}
-          />
-
-          <NavbarItem style={{ marginRight: 'auto' }}>
-            <BrandItem title={title} />
-          </NavbarItem>
-        </NavbarBrand>
+        <TabletNavbarBrand isOpen={isOpen} onMenuClick={onMenuClick} title={title} />
 
         <NavbarMenu
           style={{ justifyContent: 'center' }}
@@ -96,42 +41,29 @@ const Navigation = ({ onMenuClick, isOpen, categoryNodes, title }) => {
           onClick={onMenuClick}
         >
           <StyledNavbarItem text="About" />
-          <NavbarItem hasDropdown isHoverable>
-            <NavbarLink href="/products" style={{ paddingRight: '2em' }}>
-              Products
-            </NavbarLink>
-            <NavbarDropdown style={{}}>
-              {categoryNodes.map(node => (
-                <NavbarItem
-                  key={node.id}
-                  href={`/product-categories/${node.fields.categoryKey}`}
-                >
-                  {node.frontmatter.name}
-                </NavbarItem>
-              ))}
-            </NavbarDropdown>
-          </NavbarItem>
-          <StyledNavbarItem text="BeautyCounter" />
 
+          <ProductsDropdownItem categoryNodes={categoryNodes} />
+
+          <StyledNavbarItem text="BeautyCounter" />
           <StyledNavbarItem text="Services" />
           <StyledNavbarItem text="Contact" />
         </NavbarMenu>
       </Navbar>
     </div>
-  )
-}
+  );
+};
 
 class NavigationContainer extends React.Component {
   constructor(props) {
-    super(props)
-    this.state = { isOpen: false }
+    super(props);
+    this.state = { isOpen: false };
   }
 
   onMenuClick = () => {
     this.setState(({ isOpen: prevIsOpen }) => ({
       isOpen: !prevIsOpen,
-    }))
-  }
+    }));
+  };
 
   render() {
     return (
@@ -140,8 +72,8 @@ class NavigationContainer extends React.Component {
         onMenuClick={this.onMenuClick}
         {...this.props}
       />
-    )
+    );
   }
 }
 
-export default NavigationContainer
+export default NavigationContainer;
